+++ lib/representable/inheritable.rb
@@ -18,29 +18,33 @@ module Representable
     class Hash < ::Hash
       include Inheritable
 
-      def inherit!(parent)
-        #merge!(parent.clone)
-        for key in (parent.keys + keys).uniq
-          next unless parent_value = parent[key]
+      module InstanceMethods
+        def inherit!(parent)
+          #merge!(parent.clone)
+          for key in (parent.keys + keys).uniq
+            next unless parent_value = parent[key]
 
-          self[key].inherit!(parent_value) and next if self[key].is_a?(Inheritable)
-          self[key] = parent_value.clone and next if parent_value.is_a?(Cloneable)
+            self[key].inherit!(parent_value) and next if self[key].is_a?(Inheritable)
+            self[key] = parent_value.clone and next if parent_value.is_a?(Cloneable)
 
-          self[key] = parent_value # merge! behaviour
+            self[key] = parent_value # merge! behaviour
+          end
+
+          self
         end
 
-        self
+        def clone
+          self.class[ collect { |k,v| [k, clone_value(v)] } ]
+        end
+
+      private
+        def clone_value(value)
+          return value.clone if value.is_a?(Cloneable)
+          value
+        end
       end
 
-      def clone
-        self.class[ collect { |k,v| [k, clone_value(v)] } ]
-      end
-
-    private
-      def clone_value(value)
-        return value.clone if value.is_a?(Cloneable)
-        value
-      end
+      include InstanceMethods
     end
   end
 end