+++ lib/representable/json.rb
@@ -11,16 +11,14 @@ module Representable
       base.class_eval do
         include Representable # either in Hero or HeroRepresentation.
         extend ClassMethods # DISCUSS: do that only for classes?
-        extend Representable::Hash::ClassMethods  # DISCUSS: this is only for .from_hash, remove in 2.3?
         register_feature Representable::JSON
       end
     end
 
 
     module ClassMethods
-      # Creates a new object from the passed JSON document.
-      def from_json(*args, &block)
-        create_represented(*args, &block).from_json(*args)
+      def collection_representer_class
+        JSON::Collection
       end
     end